ISSUE 33884
https://github.com/spring-projects/spring-framework/issues/33884

This commit add implementation on ControllerTest using MockMvc of
Spring Ecossystem, all related implementation are the same as
described in approval using cURL related on topic 1 of README.md
on root of this project. The expectations using this framework
was not archived because the behavior was different from the
realtime response execution using cURL. The analysis behavior
of application design are related on README.md topic 2.